chore: add docs and refactor noise.
