--- types/src/jsonrpc_v2/error.rs
@@ -1,10 +1,0 @@
-use thiserror::Error;
-
-/// Error.
-#[derive(Error, Debug)]
-pub enum RpcError {
-	#[error("unknown rpc error")]
-	Unknown,
-	#[error("invalid params")]
-	InvalidParams,
-}
