+++ utils/src/lib.rs
@@ -1,3 +1,7 @@
+//! Shared utilities for `jsonrpsee`.
+
+#![warn(missing_docs)]
+
 #[cfg(all(feature = "hyper13", feature = "hyper14"))]
 compile_error!("feature `hyper13` and `hyper14` are mutably exclusive");
 
@@ -10,4 +14,8 @@ extern crate hyper13 as hyper;
 #[cfg(all(feature = "hyper14", not(feature = "hyper13")))]
 extern crate hyper14 as hyper;
 
+/// Helpers for HTTP.
 pub mod http;
+
+/// Helpers for JSON-RPC servers.
+pub mod server_utils;
