+++ ws-server/src/server/module.rs
@@ -1,6 +1,7 @@
 use crate::server::{Methods, RpcError, RpcParams, SubscriptionId, SubscriptionSink};
 use jsonrpsee_types::error::Error;
-use jsonrpsee_types::jsonrpc_v2::{helpers::send_response, traits::RpcMethod};
+use jsonrpsee_types::v2::traits::RpcMethod;
+use jsonrpsee_utils::server_utils::send_response;
 use parking_lot::Mutex;
 use rustc_hash::FxHashMap;
 use serde::Serialize;
